Change service directory permission; change python3 to python38
User inside the container images for logscraper and loggearman is
setting lowest value that is set in the system which is 1000.
This uid and gid is provided for other user and the services should
be running with different uids/gids.
In that case, the logscraper service gid/uid is 10210 and
loggearman gid/uid is set to 10211.

Change-Id: Ida0e2ceaf341fb7cbea18f3eaf161daa836e8ea7