Add Logsender tool
The logsender tool will be used to parse the content downloaded
by logscraper tool with --download parameter and send it to the
Elasticsearch service.
By using logsender tool, services like: gearman client, gearman
worker and logstash are not necessary to maintain, because
the information are send directly to the Elasticsearch host.

Depends-On: https://review.opendev.org/c/openstack/ci-log-processing/+/830169

Change-Id: I3e28b847e7cafbf4742fb718ef09ce3e658f945f