Add workaround for custom builds to be send
The logsender treads build dirs that does not contain full
information like: buildinfo and inventory.yaml file as incomplete -
the logsender can check "ready" directories in same time with logscraper
that is downloading the files.
The workaround for it is create not empty inventory.yaml file, so
the logsender will send it.

Change-Id: I4c606c79eeeea6cff8815a1573d1d9ca4829046a