Filter directories before download log files
The config files to check is growing up so the
logscraper needs to check a lot of urls if it contains
a log file. It is producing a unecessary taffic on logserver host.
In that case, let's check if there is a main directory available and
if it is, try to download only those files.

Change-Id: I4f8ba6fe5833710105a755f05047791e9718cf0e