Add new information for Opensearch cluster configuration
The elasticsearch-recheck project is doing queries to the
Opensearch cluster via python client. This library on
making the request does not fill the tenant, so it can not
authenticate to the Opensearch service and it raise an error:

    AuthorizationException(403, 'security_exception', 'no permissions for [cluster:monitor/main] (...) )

There is a dedicated role to allow all cluster:monitor/* parameters [1][2],
but in that case it should not be necessary.

[1] https://opensearch.org/docs/latest/security-plugin/access-control/default-action-groups/
[2] https://opensearch.org/docs/latest/security-plugin/access-control/permissions/#cluster-permissions

Change-Id: I5b017d992c7367187d2f59fd7588a8bdcc58009d