Add Opensearch dashboards backup script; add dashboards objects
The OpenSearch dashboards objects like index-pattern, visualizations
and dashboards should be versioned.
This commit just provide a tool to make a backup of OpenSearch
dashboards objects and verify in simple deployment, if the objects
have been restored correctly.
The whole mechanism to replace visualizations and dashboards that
currently are available in OpenSearch Dashboards will be done in
next patch set.

Change-Id: I4f14d9309ae0a943dbaafd220c88c76dba793f91