Check if Zuul API is reachable before try pulling logs
The logscraper service when Zuul API is down, it is raise an
exception and the service is going to stop, but the logscraper
service is controlled by systemd, which is periodically restarting
the service.
This patch is checking if the Zuul API is reachable before pull,
so it should not raise an exception and service should be running
for long time.

Change-Id: Idbe9b2024d0d049df26cf857a5309c3b9e42b80e