Add job with 'timed_out' status to fetch; add hosts_region info
Before that commit, losgcraper was pulling logs from job status
that was "SUCCESS" or "FAILURE". It was because job with other
statuses sometimes it does not contain important information to parse,
so the logscraper was creating "custom-job-result" to avoid situation
that logsender might fail to parse the build content.
This commit extends job to be pulled from logserver with field
"timed_out".
Also added new field into the document: "hosts_region" that provides
combination of host "cloud" and host "region" information.
It might be useful for infra team to check the infrastructure issues.

Change-Id: I25808709008f7e3a457d6cb4b73920619b292132