Remove build dirs with missing files; add more debug logs
It is very rary, but still from time to time there are some build
directories, that does not contain necessary files and can not be
proceded by the logsender.
This commit will provide a file with a timestamp, when the build
directory can not be processed by logsender and if after 12 hours nothing
changed, dir would be removed.
Also added more debug logs to see, what files are empty and skipped or
if buildinfo file is created properly.

Change-Id: Ie2fbde3bf7946a3aa84bfd60f87616afa0510033