Add capability with new Zuul; disable log-gearman CI job
The OpenDev infra has updated Zuul to new version (10.0.1 - ae258ac2e),
which has changed build info structure, so the Logsender
does not set important information.
Also disable gearman CI job, due it is difficult to fix to pass the CI
and the gearman functionality would be removed in the future.

Change-Id: Ifc1493b49cc10f4d7ab2b623511535ed9b665a77