Make some test queries do by using admin; change index output
The es_username is used by the logsender to push data to the OpenSearch.
Currently, the tests are using admin user to do that, but in the future
it will be replaced by dedidacated user, that will have limited access.
This commit is providing new variables just for the current tests to
execute command as admin user.
Also this commit also fix an issue when the es_index_prefix is set.

Change-Id: I4759f6ee009e3ef32160a79f9da742a78576161e